// packages/evm/src/utils/Api/models/HotListLicensePlate.ts
export type LicensePlateTemplate = {
    id: number;
    LicensePlate: string;
    DateOfInterest: string;
    LicenseType: string;
    LicenseYear: number;
    Agency: string;
    State: string;
    FirstName: string;
    LastName: string;
    Alias: string;
    VehicleYear: string;
    VehicleMake: string;
    VehicleModel: string;
    VehicleColor: string;
    VehicleStyle: string;
    Notes: string;
    NCICNumber: string;
    ImportSerialId: string;
    ViolationInfo: string;
  }